+ Bug: equipTonnage is double when loading XML + Bug: need flag for MiscType when checking whether equipment type is MASC

+++ MekHQ/src/mekhq/campaign/parts/equipment/EquipmentPart.java
@@ -192,7 +192,7 @@ public class EquipmentPart extends Part {
 				typeName = wn2.getTextContent();
 			}
 			else if (wn2.getNodeName().equalsIgnoreCase("equipTonnage")) {
-				equipTonnage = Integer.parseInt(wn2.getTextContent());
+				equipTonnage = Double.parseDouble(wn2.getTextContent());
 			}
 		}
 		restore();
